Update entrypoint.sh
Do not reinstall speedtest if already installed.

+++ entrypoint.sh
@@ -4,10 +4,13 @@ printenv >> /etc/environment
 ln -snf /usr/share/zoneinfo/$TZ /etc/localtime && echo $TZ > /etc/timezone
 
 # Install speedtest-cli
-apt-key adv --keyserver keyserver.ubuntu.com --recv-keys 379CE192D401AB61
-echo "deb https://ookla.bintray.com/debian buster main" | tee  /etc/apt/sources.list.d/speedtest.list
-apt-get update && apt-get -q -y install speedtest
-#apt-get -q -y autoremove && apt-get -q -y clean
-#rm -rf /var/lib/apt/lists/*
+if [ ! -e /usr/bin/speedtest ]
+then
+  apt-key adv --keyserver keyserver.ubuntu.com --recv-keys 379CE192D401AB61
+  echo "deb https://ookla.bintray.com/debian buster main" | tee  /etc/apt/sources.list.d/speedtest.list
+  apt-get update && apt-get -q -y install speedtest
+  apt-get -q -y autoremove && apt-get -q -y clean
+  rm -rf /var/lib/apt/lists/*
+fi
 
 exec /usr/local/bin/python3 $@
